refactor(speech): extract pure quit-phrase + TTS-markdown helpers from state_machine
`voice_loop/state_machine.rs` had grown to 2520 lines, well over the
~500-line guideline. The bulk that doesn't belong to the loop itself is
two clusters of pure, self-contained helpers (plus their characterisation
tests) that the state machine merely calls on the commit boundary:

- TTS markdown stripping (`strip_markdown_for_tts` + `consume_digit_times`
  + `find_paired_marker`) → new `voice_loop/tts_markdown.rs`
- locale-aware quit-phrase detection (`quit_phrases_for` + `is_quit_phrase`
  + `normalise_for_match`) → new `voice_loop/quit_detect.rs`

Both modules carry their existing tests verbatim (15 total) and expose
only the two entry points the state machine needs via `pub(super)`. This
follows the established `voice_loop/` "one cohesive concern per file"
pattern (cf. the per-builder `backends_*` modules). `state_machine.rs`
drops to 2202 lines.

Pure relocation — no behaviour change, identical function bodies and test
assertions. Gates: `cargo test --workspace` (default) 0 failures;
`cargo clippy --workspace --all-targets -D warnings` clean; `cargo fmt
--check` clean; `primer-speech --features voice-loop` 101 lib tests pass
(incl. all 15 relocated tests under their new module paths).
