+++ src/crates/primer-speech/src/voice_loop/state_machine.rs
@@ -10,30 +10,8 @@ use std::path::PathBuf;
 use primer_core::error::Result;
 
 use super::observer::{ExitReason, LoopObserver, TurnCompletePayload, VoiceState};
-
-/// Per-locale quit phrases. If heard in the child's transcript, the
-/// session ends. Case-insensitive, word-boundary match (see
-/// [`is_quit_phrase`]). Each locale ships its own set so a child can
-/// quit in the language they speak — without a locale-aware list, the
-/// German voice mode silently lacks any voice-keyword end affordance.
-///
-/// Adding a new locale: append a `(pack_id, &[phrase, ...])` row. The
-/// pack_id must match `Locale::pack_id()` for the corresponding locale.
-fn quit_phrases_for(locale: &primer_core::i18n::Locale) -> &'static [&'static str] {
-    match locale.pack_id() {
-        "de" => &[
-            // "Tschüss" (informal goodbye) — the most natural for a child.
-            "tschüss",
-            // Formal goodbye.
-            "auf wiedersehen",
-            // Primer-direct variants, mirroring the EN set.
-            "bye primer",
-            "stop primer",
-        ],
-        // English is the default for any unrecognised locale.
-        _ => &["goodbye", "bye primer", "stop primer"],
-    }
-}
+use super::quit_detect::is_quit_phrase;
+use super::tts_markdown::strip_markdown_for_tts;
 
 /// Spoken when the LLM call fails (rate limit, network, etc.). Goes
 /// through Piper just like any normal Primer turn — the child hears
@@ -75,157 +53,6 @@ fn handle_llm_err<O: LoopObserver>(
     chunks.clear();
     chunks.push(FALLBACK_LINE.to_string());
     FALLBACK_LINE.to_string()
-}
-
-/// Returns true if `transcript`, after trimming surrounding whitespace
-/// and punctuation, **equals** one of `locale`'s quit phrases
-/// (case-insensitive).
-///
-/// Why exact-equality rather than `contains` or word-boundary: the
-/// pre-fix `contains` would end the session on *"I don't want to stop
-/// primer"* because the substring `"stop primer"` matched. Word-boundary
-/// matching alone doesn't help — end-of-string is itself a word boundary.
-/// The only safe contract for an auto-quit voice keyword is "the child
-/// said exactly the keyword, nothing else." Children can always say
-/// "goodbye" by itself; this also matches the way real children end
-/// conversations.
-///
-/// Trimming punctuation handles Whisper's habit of producing trailing
-/// `.` / `!` / `?` on a finalized utterance: `"Goodbye!"` still ends the
-/// session. Internal whitespace is normalised to single spaces so the
-/// transcript `"bye   primer"` matches `"bye primer"`.
-fn is_quit_phrase(transcript: &str, locale: &primer_core::i18n::Locale) -> bool {
-    let normalised = normalise_for_match(transcript);
-    quit_phrases_for(locale)
-        .iter()
-        .any(|p| normalised == normalise_for_match(p))
-}
-
-/// Lowercase, strip leading/trailing whitespace + punctuation, collapse
-/// internal whitespace to single spaces. The result is the canonical
-/// form used for quit-phrase equality.
-///
-/// `char::is_alphanumeric` is Unicode-aware so German `ü`/`ö`/`ä` are
-/// preserved; only punctuation and non-letter symbols get trimmed.
-fn normalise_for_match(s: &str) -> String {
-    let lower = s.to_lowercase();
-    let trimmed = lower.trim_matches(|c: char| !c.is_alphanumeric());
-    let mut out = String::with_capacity(trimmed.len());
-    let mut prev_was_space = false;
-    for c in trimmed.chars() {
-        if c.is_whitespace() {
-            if !prev_was_space {
-                out.push(' ');
-                prev_was_space = true;
-            }
-        } else {
-            out.push(c);
-            prev_was_space = false;
-        }
-    }
-    out
-}
-
-/// Strip markdown emphasis markers so Piper's espeak phonemizer doesn't
-/// pronounce them ("*why*" → "asterisks why asterisks"). Paired
-/// `*emphasis*` and `**strong**` markers are removed; paired
-/// `` `code` `` markers are removed. Bare unmatched `*` or `` ` `` are
-/// left in place. A `*` (or run of `*`) sandwiched between digits is
-/// treated as multiplication and replaced with " times " so `5*3=15`
-/// reads as "5 times 3=15" instead of "53=15". Underscore-emphasis is
-/// rare and ambiguous (shows up in identifiers too) — left alone.
-///
-/// Recursion: the function recurses into the inner content of paired
-/// markers (e.g. the `5*3=15` inside `**5*3=15**`). Each recursive call
-/// receives a strict substring, so depth is bounded by `input.len()/2`
-/// and stack overflow is impossible for any realistic Primer turn.
-fn strip_markdown_for_tts(text: &str) -> String {
-    let chars: Vec<char> = text.chars().collect();
-    let mut out = String::with_capacity(text.len());
-    let mut i = 0;
-    while i < chars.len() {
-        let c = chars[i];
-        if c == '*' {
-            if let Some(end) = consume_digit_times(&chars, i) {
-                out.push_str(" times ");
-                i = end;
-                continue;
-            }
-            let marker = if i + 1 < chars.len() && chars[i + 1] == '*' {
-                2
-            } else {
-                1
-            };
-            if let Some(close) = find_paired_marker(&chars, i + marker, marker, '*') {
-                let inner: String = chars[i + marker..close].iter().collect();
-                out.push_str(&strip_markdown_for_tts(&inner));
-                i = close + marker;
-                continue;
-            }
-            out.push('*');
-            i += 1;
-        } else if c == '`' {
-            if let Some(close) = find_paired_marker(&chars, i + 1, 1, '`') {
-                let inner: String = chars[i + 1..close].iter().collect();
-                out.push_str(&strip_markdown_for_tts(&inner));
-                i = close + 1;
-                continue;
-            }
-            out.push('`');
-            i += 1;
-        } else {
-            out.push(c);
-            i += 1;
-        }
-    }
-    out
-}
-
-/// If `chars[i]` is the start of a digit-bounded run of `*` (e.g. `*`
-/// or `**` flanked by digits), return the index just past the run.
-/// Otherwise return `None`. The caller emits " times " in that case.
-///
-/// Only ASCII integer boundaries match: `1.5*2`, `1,000*5`, and any
-/// non-ASCII numeral won't trigger the rewrite. This is the right
-/// trade-off for a children's tutor (integer multiplication dominates),
-/// and keeps the heuristic narrow enough that it never fires on prose.
-fn consume_digit_times(chars: &[char], i: usize) -> Option<usize> {
-    if i == 0 || !chars[i - 1].is_ascii_digit() {
-        return None;
-    }
-    let mut j = i;
-    while j < chars.len() && chars[j] == '*' {
-        j += 1;
-    }
-    if j < chars.len() && chars[j].is_ascii_digit() {
-        Some(j)
-    } else {
-        None
-    }
-}
-
-/// Find the next run of exactly `marker_len` consecutive `marker`
-/// characters starting at or after `start`, not adjacent to another
-/// `marker` (so a `*` inside a `**` run never matches a single-`*`
-/// search and vice versa). Returns the start index of that run.
-fn find_paired_marker(
-    chars: &[char],
-    start: usize,
-    marker_len: usize,
-    marker: char,
-) -> Option<usize> {
-    let n = chars.len();
-    let mut i = start;
-    while i + marker_len <= n {
-        let matches = (0..marker_len).all(|k| chars[i + k] == marker);
-        let prev_ok = i == 0 || chars[i - 1] != marker;
-        let next_ok = i + marker_len >= n || chars[i + marker_len] != marker;
-        if matches && prev_ok && next_ok {
-            return Some(i);
-        }
-        i += 1;
-    }
-    None
 }
 
 /// Configuration passed into [`run_loop`] / the higher-level `run` entry
@@ -2371,150 +2198,5 @@ mod mocks {
             "must distinguish voice-profile miss from TTS miss: {msg}"
         );
         assert!(msg.contains("'en'"), "must name the missing locale: {msg}");
-    }
-}
-
-#[cfg(test)]
-mod markdown_tests {
-    use super::strip_markdown_for_tts;
-
-    #[test]
-    fn strips_paired_emphasis_and_strong() {
-        assert_eq!(strip_markdown_for_tts("*why*"), "why");
-        assert_eq!(strip_markdown_for_tts("**important**"), "important");
-        assert_eq!(
-            strip_markdown_for_tts("a *little* bit of **emphasis**"),
-            "a little bit of emphasis"
-        );
-    }
-
-    #[test]
-    fn preserves_multiplication_between_digits() {
-        assert_eq!(strip_markdown_for_tts("5*3=15"), "5 times 3=15");
-        assert_eq!(strip_markdown_for_tts("2 * 3"), "2 * 3");
-        assert_eq!(strip_markdown_for_tts("5*3*2"), "5 times 3 times 2");
-    }
-
-    #[test]
-    fn preserves_exponent_double_star_between_digits() {
-        assert_eq!(strip_markdown_for_tts("5**2"), "5 times 2");
-    }
-
-    #[test]
-    fn leaves_unmatched_star_alone() {
-        assert_eq!(strip_markdown_for_tts("a* footnote"), "a* footnote");
-        assert_eq!(strip_markdown_for_tts("value *= 5"), "value *= 5");
-    }
-
-    #[test]
-    fn strips_paired_backticks_only() {
-        assert_eq!(strip_markdown_for_tts("`code`"), "code");
-        assert_eq!(
-            strip_markdown_for_tts("a single ` backtick"),
-            "a single ` backtick"
-        );
-    }
-
-    #[test]
-    fn handles_mixed_markdown_and_math() {
-        assert_eq!(
-            strip_markdown_for_tts("the answer is **5*3=15** indeed"),
-            "the answer is 5 times 3=15 indeed"
-        );
-    }
-
-    #[test]
-    fn no_op_on_plain_text() {
-        assert_eq!(
-            strip_markdown_for_tts("nothing to strip here"),
-            "nothing to strip here"
-        );
-    }
-
-    #[test]
-    fn empty_input_returns_empty() {
-        assert_eq!(strip_markdown_for_tts(""), "");
-    }
-
-    /// Triple-`*` runs (bold-italic markdown) are not currently
-    /// recognised — the inner closer is rejected by the
-    /// "not adjacent to another marker" guard and the outer pair
-    /// finds no match. Pinned here so a future refactor doesn't
-    /// silently break the current behaviour. If this assertion
-    /// ever needs updating, re-derive the right output from first
-    /// principles rather than tweaking the test.
-    #[test]
-    fn triple_star_passes_through_unchanged_for_now() {
-        assert_eq!(strip_markdown_for_tts("***foo***"), "***foo***");
-    }
-}
-
-#[cfg(test)]
-mod quit_tests {
-    use super::is_quit_phrase;
-    use primer_core::i18n::Locale;
-
-    #[test]
-    fn detects_goodbye_case_insensitive() {
-        assert!(is_quit_phrase("Goodbye!", &Locale::English));
-        assert!(is_quit_phrase("GOODBYE", &Locale::English));
-    }
-
-    #[test]
-    fn detects_bye_primer() {
-        assert!(is_quit_phrase("bye primer", &Locale::English));
-        assert!(is_quit_phrase("Bye Primer.", &Locale::English));
-    }
-
-    #[test]
-    fn ignores_unrelated_transcripts() {
-        assert!(!is_quit_phrase("why is the sky blue", &Locale::English));
-        assert!(!is_quit_phrase("hello", &Locale::English));
-        // "bye" alone is NOT a quit phrase — only "bye primer".
-        assert!(!is_quit_phrase("bye", &Locale::English));
-    }
-
-    /// Embedded-phrase guard: a quit phrase embedded inside a longer
-    /// utterance must NOT terminate the session. The pre-fix substring
-    /// `contains` would have ended the session on either of these —
-    /// exactly the opposite of the child's intent. Word-boundary
-    /// matching alone wouldn't fix this (end-of-string is itself a
-    /// word boundary); equality-after-normalisation does.
-    #[test]
-    fn embedded_quit_phrase_does_not_end_session() {
-        assert!(!is_quit_phrase(
-            "I don't want to stop primer",
-            &Locale::English
-        ));
-        assert!(!is_quit_phrase("alright goodbye then", &Locale::English));
-        // ... but the phrase as a complete utterance ends the session.
-        assert!(is_quit_phrase("stop primer", &Locale::English));
-        // Punctuation around the phrase is fine (Whisper often appends).
-        assert!(is_quit_phrase("Stop primer!", &Locale::English));
-        // Collapsed internal whitespace is fine too.
-        assert!(is_quit_phrase("  bye   primer  ", &Locale::English));
-    }
-
-    /// German locale ships its own quit phrases. A German-speaking child
-    /// who says "tschüss" or "auf wiedersehen" must be able to end the
-    /// session by voice — and an English "goodbye" should NOT match in
-    /// a German session.
-    #[test]
-    fn german_locale_uses_german_quit_phrases() {
-        assert!(is_quit_phrase("tschüss", &Locale::German));
-        assert!(is_quit_phrase("Tschüss!", &Locale::German));
-        assert!(is_quit_phrase("auf wiedersehen", &Locale::German));
-        assert!(is_quit_phrase("Auf Wiedersehen.", &Locale::German));
-        // English-only phrases don't end a German session.
-        assert!(!is_quit_phrase("goodbye", &Locale::German));
-        // Primer-direct variants are universal (English loanwords).
-        assert!(is_quit_phrase("bye primer", &Locale::German));
-    }
-
-    /// English locale must NOT match German-only phrases.
-    #[test]
-    fn english_locale_rejects_german_only_phrases() {
-        assert!(!is_quit_phrase("tschüss", &Locale::English));
-        assert!(!is_quit_phrase("auf wiedersehen", &Locale::English));
     }
 }
